+++ packages/@expo/metro-config/src/transform-worker/transform-worker.ts
@@ -41,6 +41,30 @@ export async function transform(
   const isCss = options.type !== 'asset' && /\.(s?css|sass)$/.test(filename);
   // If the file is not CSS, then use the default behavior.
   if (!isCss) {
+    const environment = options.customTransformOptions?.environment;
+
+    if (
+      environment === 'client' &&
+      // TODO: Ensure this works with windows.
+      // TODO: Add +api files.
+      filename.match(new RegExp(`^app/\\+html(\\.${options.platform})?\\.([tj]sx?|[cm]js)?$`))
+    ) {
+      // Remove the server-only +html file from the bundle when bundling for a client environment.
+      return worker.transform(
+        config,
+        projectRoot,
+        filename,
+        options.dev
+          ? Buffer.from(
+              // Use a string so this notice is visible in the bundle if the user is
+              // looking for it.
+              '"> The server-only +html file was removed from the client JS bundle by Expo CLI."'
+            )
+          : Buffer.from(''),
+        options
+      );
+    }
+
     return worker.transform(config, projectRoot, filename, data, options);
   }